feat: HTML scaffolding, session storage, and fixer.io api interactions
HTML scaffolding
Created view for basic page. Should be web accessible.

Session Storage
Storing the currency conversion as a new item. Get's the item before
page load.

Fixer.io API interactions
Using the browsers fetch API, a promise object is made explicitly
defining the USD currency to be the parent. A promise is created and
returns a json response. Grabbing the GBP object to as the correct
conversion and rewrites the anchor tag with the new value.